change pc passed to backend into 32bit

+++ accel/tcg/tcg-runtime-sym.c
@@ -651,7 +651,7 @@ void HELPER(sym_notify_return)(uint64_t return_address)
     _sym_notify_ret(return_address);
 }
 
-void HELPER(sym_notify_block)(uint64_t block_id)
+void HELPER(sym_notify_block)(uint32_t block_id)
 {
     _sym_notify_basic_block(block_id);
 }
