+++ accel/tcg/tcg-runtime-sym.h
@@ -119,7 +119,7 @@ DEF_HELPER_FLAGS_7(sym_setcond_i64, TCG_CALL_NO_RWG, ptr, env, i64, ptr, i64, pt
 /* Context tracking */
 DEF_HELPER_FLAGS_2(sym_notify_call, TCG_CALL_NO_RWG, void, i64, i64)
 DEF_HELPER_FLAGS_1(sym_notify_return, TCG_CALL_NO_RWG, void, i64)
-DEF_HELPER_FLAGS_1(sym_notify_block, TCG_CALL_NO_RWG, void, i64)
+DEF_HELPER_FLAGS_1(sym_notify_block, TCG_CALL_NO_RWG, void, i32)
 
 /* indirect jmp and indirect calls */
 // DEF_HELPER_FLAGS_1(sym_notify_jmp, TCG_CALL_NO_RWG, void, i64)
